# openapi.yaml
openapi: 3.0.0
info:
  title: AFOCE Accounting API
  description: |
    Complete accounting API for invoicing, expenses, and financial management.
    Supports multi-tenant workspaces with role-based access control.
  version: 1.0.0
  contact:
    name: AFOCE Support
    email: support@afoce.app
servers:
  - url: https://api.afoce.app
    description: Production server
  - url: http://localhost:3000
    description: Local development
security:
  - cookieAuth: []
paths:
...
  /api/upload:
    post:
      summary: Upload file
      operationId: uploadFile
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                type:
                  type: string
                  enum: [receipt, avatar, invoice, general]
      responses:
        "200":
          description: File uploaded
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: supabase-auth-token
...